Refactory: S3 DTO 수정
modify: dto 파일 및 폴더 이름 변경
delete: s3 controller 삭제 ← 사용하지 않는데 남겨둘 필요 없음
