+++ src/s3/dto/aws-s3-copy-file.dto.ts
@@ -1,4 +1,4 @@
-import { AwsS3DeleteFileDto } from './aws-s3-delete-files.dto';
+import { AwsS3DeleteFileDto } from './aws-s3-delete-file.dto';
 
 export interface AwsS3CopyFileDto extends AwsS3DeleteFileDto {
   destinationKey: string;
