+++ src/s3/dto/aws-s3-delete-file.dto.ts
