+++ src/s3/dto/aws-s3-move-file.dto.ts
@@ -1,3 +1,3 @@
-import { AwsS3CopyFileDto } from './aws-s3-copy-files.dto';
+import { AwsS3CopyFileDto } from './aws-s3-copy-file.dto';
 
 export interface AwsS3MoveFileDto extends AwsS3CopyFileDto {}
